aFewStepsFoward

+++ levels.c
@@ -25,8 +25,8 @@ void level1(breakable_t *breakable, ball_t *ball, player_t *player) {
     //Ball 1
     ball[0].whatIsThis = 0x02;
     ball[0].changedSinceLast = 1;
-    ball[0].xs = LONG_TO_EIGHTEEN_FOURTEEN(40);
-    ball[0].ys = LONG_TO_EIGHTEEN_FOURTEEN(25);
+    ball[0].xs = LONG_TO_EIGHTEEN_FOURTEEN((long)40);
+    ball[0].ys = LONG_TO_EIGHTEEN_FOURTEEN((long)25);
     ball[0].x1 = ball[0].xs;
     ball[0].y1 = ball[0].ys;
     ball[0].direction = tempVec;
@@ -34,8 +34,8 @@ void level1(breakable_t *breakable, ball_t *ball, player_t *player) {
     //Ball 2
     ball[1].whatIsThis = 0x02;
     ball[1].changedSinceLast = 1;
-    ball[1].xs = LONG_TO_EIGHTEEN_FOURTEEN(45);
-    ball[1].ys = LONG_TO_EIGHTEEN_FOURTEEN(25);
+    ball[1].xs = LONG_TO_EIGHTEEN_FOURTEEN((long)45);
+    ball[1].ys = LONG_TO_EIGHTEEN_FOURTEEN((long)25);
     ball[1].x1 = ball[1].xs;
     ball[1].y1 = ball[1].ys;
     ball[1].direction = tempVec;
@@ -43,8 +43,8 @@ void level1(breakable_t *breakable, ball_t *ball, player_t *player) {
     //Ball 3
     ball[2].whatIsThis = 0x02;
     ball[2].changedSinceLast = 1;
-    ball[2].xs = LONG_TO_EIGHTEEN_FOURTEEN(50);
-    ball[2].ys = LONG_TO_EIGHTEEN_FOURTEEN(25);
+    ball[2].xs = LONG_TO_EIGHTEEN_FOURTEEN((long)50);
+    ball[2].ys = LONG_TO_EIGHTEEN_FOURTEEN((long)25);
     ball[2].x1 = ball[2].xs;
     ball[2].y1 = ball[2].ys;
     ball[2].direction = tempVec;
@@ -52,8 +52,8 @@ void level1(breakable_t *breakable, ball_t *ball, player_t *player) {
     //Ball 4
     ball[3].whatIsThis = 0x02;
     ball[3].changedSinceLast = 1;
-    ball[3].xs = LONG_TO_EIGHTEEN_FOURTEEN(55);
-    ball[3].ys = LONG_TO_EIGHTEEN_FOURTEEN(25);
+    ball[3].xs = LONG_TO_EIGHTEEN_FOURTEEN((long)55);
+    ball[3].ys = LONG_TO_EIGHTEEN_FOURTEEN((long)25);
     ball[3].x1 = ball[3].xs;
     ball[3].y1 = ball[3].ys;
     ball[3].direction = tempVec;
